LayoutTests/pointer-lock/mouse-event-api-expected.txt

PASS mouseEvent.movementX is defined.
PASS mouseEvent.movementY is defined.
PASS successfullyParsed is true

TEST COMPLETE
